pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "Hypercrystal"
version = "0.1.5"
dependencies = [
  "pygame-ce"
]
requires-python = ">=3.12"
authors = [
  {name = "Nix"},
]
maintainers = []
description = "Python framework for realtime hyperbolic rendering in H2, build as a layer above pygame-ce."
readme = "package_readme.md"
license = { file = "LICENSE" }
license-files = ["LICEN[CS]E.*"]
keywords = []
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "License :: OSI Approved :: MIT License",
]

[project.urls]
Author = "https://www.youtube.com/@tucan7112"
Github = "https://github.com/Tucan444/Hypercrystal"

[tool.hatch.build.targets.wheel]
packages = ["src/hypercrystal"]

[tool.hatch.build.targets.sdist]
include = [
  "src/hypercrystal",
  "package_readme.md",
  "LICENSE",
]
